refactor: more menus
