+++ packages/shared/src/components/molecules/menu-items/index.js
@@ -1,2 +1,1 @@
 export { default as AccountSwitcherMenuItem } from './AccountSwitcherMenuItem.svelte'
-export { default as ToggleHiddenAccountMenuItem } from './ToggleHiddenAccountMenuItem.svelte'
